requirements.txt
PyQt6>=6.6.0
anthropic>=0.20.0
python-dotenv>=1.0.0
Pillow>=10.0.0
